HTTP API doc generation
