+++ app.js
@@ -7,6 +7,8 @@ const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
 const fs = require('fs');
+const bodyParser = require('body-parser');
+
 
 const app = express();
 
@@ -37,15 +39,6 @@ app.models.sequelize.sync().then(() => {
   });
 });
 
-
-const usersRouter = require('./routes/users')(app);
-const bansRouter = require('./routes/bans')(app);
-const pagesRouter = require('./routes/pages')(app);
-
-app.use('/', pagesRouter);
-app.use('/user', usersRouter);
-app.use('/ban', bansRouter);
-
 // view engine setup
 app.set('view engine', 'pug');
 app.set('views', path.join(__dirname, 'views'));
@@ -58,10 +51,11 @@ app.use(express.json());
 app.use(express.urlencoded({
   extended: false,
 }));
+app.use(bodyParser.json());
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-
+require('./routes')(app);
 // catch 404 and forward to error handler
 app.use((req, res, next) => {
   next(createError(404, 'This page does not exist.'));
@@ -78,7 +72,6 @@ app.use((err, req, res) => {
   res.status(err.status || 500);
   res.render('error');
 });
-
 
 // Register helper functions
 app.helpers = {};
