+++ bin/www
@@ -5,7 +5,6 @@
  */
 
 var app = require('../app');
-var debug = require('debug')('global-ban-list:server');
 var http = require('http');
