+++ package.json
@@ -13,6 +13,7 @@
     "docs:api": "./node_modules/.bin/apidoc -i routes/ -o docs/api/"
   },
   "dependencies": {
+    "body-parser": "^1.18.3",
     "cookie-parser": "~1.4.3",
     "debug": "~2.6.9",
     "dotenv": "^6.1.0",
